issues fixed

+++ src/main/java/ru/skypro/lessons/springboot/weblibrary/controller/EmployeeController.java
@@ -27,7 +27,7 @@ public class EmployeeController {
     }
 
     @GetMapping("salary/sum")
-    public long findSumSalariesOfEmployees() {
+    public double findSumSalariesOfEmployees() {
         return employeeService.findSumSalariesOfEmployees();
     }
