+++ src/main/java/ru/skypro/lessons/springboot/weblibrary/service/EmployeeService.java
@@ -8,7 +8,7 @@ public interface EmployeeService {
     List<Employee> getAllEmployees();
     Employee findEmployeeWithMinSalary();
     Employee findEmployeeWithMaxSalary();
-    long findSumSalariesOfEmployees();
+    double findSumSalariesOfEmployees();
     List<Employee> findEmployeesWithAboveAverageSalaries();
 
 }
